+++ src/ios/frameworks/OSCameraLib.xcframework/ios-arm64/OSCameraLib.framework/_CodeSignature/CodeResources
@@ -18,15 +18,15 @@
 		</data>
 		<key>Modules/OSCameraLib.swiftmodule/arm64-apple-ios.swiftdoc</key>
 		<data>
-		xuVAlcAoFLNuaBDG4gD184LDuNQ=
+		onsuf/lITrOUgp16x7X6Z1sHu2Y=
 		</data>
 		<key>Modules/OSCameraLib.swiftmodule/arm64-apple-ios.swiftinterface</key>
 		<data>
-		r4b9zBBCyYCompF/sCe/5uCSBZo=
+		buq7rKbIzJKZ6UUimejTyCfLg5Q=
 		</data>
 		<key>Modules/OSCameraLib.swiftmodule/arm64-apple-ios.swiftmodule</key>
 		<data>
-		ASVDF3tpGj0hs7wC+y2usCsmkTw=
+		6CBIpCyBbWCSqvVDCOLStcgN2+4=
 		</data>
 		<key>Modules/module.modulemap</key>
 		<data>
@@ -53,21 +53,21 @@
 		<dict>
 			<key>hash2</key>
 			<data>
-			VZ8PtyENyKFT/w0llWbeeUDz7HZLHJsSFLLJg176AfM=
+			SILoAhWeWeDJGyynwivna2z9ER6TD6Y0xdWHO4jo5WU=
 			</data>
 		</dict>
 		<key>Modules/OSCameraLib.swiftmodule/arm64-apple-ios.swiftinterface</key>
 		<dict>
 			<key>hash2</key>
 			<data>
-			MajwDGdw1dRcG2xaD/fPv5VpqWgToQ8k9yCC9e3gG1w=
+			iBKcxKUBjcK/1tB3E0YOZko2CyENnOgLYjQoYdVrQ0o=
 			</data>
 		</dict>
 		<key>Modules/OSCameraLib.swiftmodule/arm64-apple-ios.swiftmodule</key>
 		<dict>
 			<key>hash2</key>
 			<data>
-			M+cdgHYC1AumPYxuGUGLGmOjOgga9jNcWvdX89zVYFw=
+			liskVKe8yr8EvRnABa1kKTSFxfnL8uT0YLoNKrC8PYE=
 			</data>
 		</dict>
 		<key>Modules/module.modulemap</key>
